api/requirements.txt
APIFlask==2.2.1
pymongo==4.9.1
python-dateutil==2.9.0.post0
requests==2.32.3
